Minor Changes to Update, Improve Aesthetics
Change processing to extract lowest level of a nested random variable to
use as the minimum unit for graphing in Model Checks JB.R

Use try() when loading HLMdiag package so code continues if package is not
installed.

Change x and y axis scaling to improve aesthetics and change panel.margin
to panel.spacing in ggplot2 code for compliance with latest package
update.