Improve Names in Graphs, R-Squared in Checks
Modifies code in Graph GLMMs.R to parse the terms present in an interaction and
create a graphing name using the names given for graphing to the individual
variables. This also standardizes the order of terms in the interaction.

Also modifies r-squared call in Model Checks JB.R to work with piecewiseSEM package
versions before or after 2.0.  Also adds some code for LRT but it isn't working yet
so is commented out.